fix 리뷰수정 #jinny

+++ TogeDutch/src/main/java/com/proj/togedutch/entity/ReviewEmotion.java
@@ -5,12 +5,13 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.sql.Timestamp;
+import java.util.List;
 
 @Getter
 @Setter
 @AllArgsConstructor
 public class ReviewEmotion {
     private int post_id;
-    private Double avg;
+    private int avg;
 
 }
